+++ lib/shard/src/update/points/sync.rs
@@ -1,0 +1,173 @@
+//! Point syncs: replace a point range with the given set, plain and raw.
+
+use std::sync::atomic::AtomicBool;
+
+use ahash::{AHashMap, AHashSet};
+use common::counter::hardware_counter::HardwareCounterCell;
+use common::types::DeferredBehavior;
+use segment::common::operation_error::OperationResult;
+use segment::types::{PointIdType, SeqNumberType, WithPayload, WithVector};
+
+use super::{delete_points, upsert_points, upsert_points_raw};
+use crate::operations::point_ops::{PointStructPersisted, PointStructRawPersisted};
+use crate::segment_holder::SegmentHolder;
+
+/// Sync points within a given [from_id; to_id) range.
+///
+/// 1. Retrieve existing points for a range
+/// 2. Remove points, which are not present in the sync operation
+/// 3. Retrieve overlapping points, detect which one of them are changed
+/// 4. Select new points
+/// 5. Upsert points which differ from the stored ones
+///
+/// Returns:
+///     (number of deleted points, number of new points, number of updated points)
+pub fn sync_points(
+    segments: &SegmentHolder,
+    op_num: SeqNumberType,
+    from_id: Option<PointIdType>,
+    to_id: Option<PointIdType>,
+    points: &[PointStructPersisted],
+    hw_counter: &HardwareCounterCell,
+) -> OperationResult<(usize, usize, usize)> {
+    let id_to_point: AHashMap<PointIdType, _> = points.iter().map(|p| (p.id, p)).collect();
+    let sync_points: AHashSet<_> = points.iter().map(|p| p.id).collect();
+    // 1. Retrieve existing points for a range
+    let stored_point_ids: AHashSet<_> = segments
+        .iter()
+        .flat_map(|(_, segment)| segment.get().read().read_range(from_id, to_id))
+        .collect();
+    // 2. Remove points, which are not present in the sync operation
+    let points_to_remove: Vec<_> = stored_point_ids.difference(&sync_points).copied().collect();
+    let deleted = delete_points(segments, op_num, points_to_remove.as_slice(), hw_counter)?;
+    // 3. Retrieve overlapping points, detect which one of them are changed
+    let existing_point_ids: Vec<_> = stored_point_ids
+        .intersection(&sync_points)
+        .copied()
+        .collect();
+
+    let mut points_to_update: Vec<_> = Vec::new();
+    // we don’t want to cancel this filtered read
+    let is_stopped = AtomicBool::new(false);
+    let _num_updated = segments.read_points(
+        existing_point_ids.as_slice(),
+        &is_stopped,
+        DeferredBehavior::WithDeferred,
+        |ids, segment| {
+            let with_vector = WithVector::Bool(true);
+            let with_payload = WithPayload::from(true);
+            // Since we retrieve points, which we already know exist, we expect all of them to be found
+            let stored_records = segment.retrieve(
+                ids,
+                &with_payload,
+                &with_vector,
+                hw_counter,
+                &is_stopped,
+                DeferredBehavior::WithDeferred,
+            )?;
+            let mut updated = 0;
+
+            for (id, stored_record) in stored_records {
+                let point = id_to_point.get(&id).unwrap();
+                if !point.is_equal_to(&stored_record) {
+                    points_to_update.push(*point);
+                    updated += 1;
+                }
+            }
+
+            Ok(updated)
+        },
+    )?;
+
+    // 4. Select new points
+    let num_updated = points_to_update.len();
+    let mut num_new = 0;
+    sync_points.difference(&stored_point_ids).for_each(|id| {
+        num_new += 1;
+        points_to_update.push(*id_to_point.get(id).unwrap());
+    });
+
+    // 5. Upsert points which differ from the stored ones
+    let num_replaced = upsert_points(segments, op_num, points_to_update, hw_counter)?;
+    debug_assert!(
+        num_replaced <= num_updated,
+        "number of replaced points cannot be greater than points to update ({num_replaced} <= {num_updated})",
+    );
+
+    Ok((deleted, num_new, num_updated))
+}
+
+pub fn sync_points_raw(
+    segments: &SegmentHolder,
+    op_num: SeqNumberType,
+    from_id: Option<PointIdType>,
+    to_id: Option<PointIdType>,
+    points: &[PointStructRawPersisted],
+    hw_counter: &HardwareCounterCell,
+) -> OperationResult<(usize, usize, usize)> {
+    let id_to_point: AHashMap<PointIdType, _> = points.iter().map(|p| (p.id, p)).collect();
+    let sync_points: AHashSet<_> = points.iter().map(|p| p.id).collect();
+    // 1. Retrieve existing points for a range
+    let stored_point_ids: AHashSet<_> = segments
+        .iter()
+        .flat_map(|(_, segment)| segment.get().read().read_range(from_id, to_id))
+        .collect();
+    // 2. Remove points, which are not present in the sync operation
+    let points_to_remove: Vec<_> = stored_point_ids.difference(&sync_points).copied().collect();
+    let deleted = delete_points(segments, op_num, points_to_remove.as_slice(), hw_counter)?;
+    // 3. Retrieve overlapping points, detect which one of them are changed
+    let existing_point_ids: Vec<_> = stored_point_ids
+        .intersection(&sync_points)
+        .copied()
+        .collect();
+
+    let mut points_to_update: Vec<_> = Vec::new();
+    // we don’t want to cancel this filtered read
+    let is_stopped = AtomicBool::new(false);
+    let _num_updated = segments.read_points(
+        existing_point_ids.as_slice(),
+        &is_stopped,
+        DeferredBehavior::WithDeferred,
+        |ids, segment| {
+            let with_vector = WithVector::Bool(true);
+            let with_payload = WithPayload::from(true);
+            // Since we retrieve points, which we already know exist, we expect all of them to be found
+            let stored_records = segment.retrieve_raw(
+                ids,
+                &with_payload,
+                &with_vector,
+                hw_counter,
+                &is_stopped,
+                DeferredBehavior::WithDeferred,
+            )?;
+            let mut updated = 0;
+
+            for (id, stored_record) in stored_records {
+                let point = id_to_point.get(&id).unwrap();
+                if !point.is_equal_to(&stored_record) {
+                    points_to_update.push(*point);
+                    updated += 1;
+                }
+            }
+
+            Ok(updated)
+        },
+    )?;
+
+    // 4. Select new points
+    let num_updated = points_to_update.len();
+    let mut num_new = 0;
+    sync_points.difference(&stored_point_ids).for_each(|id| {
+        num_new += 1;
+        points_to_update.push(*id_to_point.get(id).unwrap());
+    });
+
+    // 5. Upsert points which differ from the stored ones
+    let num_replaced = upsert_points_raw(segments, op_num, points_to_update, hw_counter)?;
+    debug_assert!(
+        num_replaced <= num_updated,
+        "number of replaced points cannot be greater than points to update ({num_replaced} <= {num_updated})",
+    );
+
+    Ok((deleted, num_new, num_updated))
+}
